Rework the net server to have a base server that can be reused to TCP based servers
