+++ src/main/java/io/vertx/core/impl/VertxImpl.java
@@ -51,6 +51,7 @@ import io.vertx.core.net.NetServer;
 import io.vertx.core.net.NetServerOptions;
 import io.vertx.core.net.impl.NetClientImpl;
 import io.vertx.core.net.impl.NetServerImpl;
+import io.vertx.core.net.impl.NetServerBase;
 import io.vertx.core.net.impl.ServerID;
 import io.vertx.core.shareddata.SharedData;
 import io.vertx.core.shareddata.impl.SharedDataImpl;
@@ -96,7 +97,7 @@ public class VertxImpl implements VertxInternal, MetricsProvider {
   private final DeploymentManager deploymentManager;
   private final FileResolver fileResolver;
   private final Map<ServerID, HttpServerImpl> sharedHttpServers = new HashMap<>();
-  private final Map<ServerID, NetServerImpl> sharedNetServers = new HashMap<>();
+  private final Map<ServerID, NetServerBase> sharedNetServers = new HashMap<>();
   private final WorkerPool workerPool;
   private final WorkerPool internalBlockingPool;
   private final ThreadFactory eventLoopThreadFactory;
@@ -320,7 +321,7 @@ public class VertxImpl implements VertxInternal, MetricsProvider {
     return sharedHttpServers;
   }
 
-  public Map<ServerID, NetServerImpl> sharedNetServers() {
+  public Map<ServerID, NetServerBase> sharedNetServers() {
     return sharedNetServers;
   }
 
@@ -488,7 +489,7 @@ public class VertxImpl implements VertxInternal, MetricsProvider {
             closeClusterManager(ar4 -> {
               // Copy set to prevent ConcurrentModificationException
               Set<HttpServer> httpServers = new HashSet<>(sharedHttpServers.values());
-              Set<NetServer> netServers = new HashSet<>(sharedNetServers.values());
+              Set<NetServerBase> netServers = new HashSet<>(sharedNetServers.values());
               sharedHttpServers.clear();
               sharedNetServers.clear();
 
@@ -508,7 +509,7 @@ public class VertxImpl implements VertxInternal, MetricsProvider {
               for (HttpServer server : httpServers) {
                 server.close(serverCloseHandler);
               }
-              for (NetServer server : netServers) {
+              for (NetServerBase server : netServers) {
                 server.close(serverCloseHandler);
               }
               if (serverCount == 0) {
