+++ src/main/java/io/vertx/core/metrics/impl/DummyVertxMetrics.java
@@ -35,7 +35,6 @@ import io.vertx.core.http.WebSocket;
 import io.vertx.core.spi.metrics.*;
 import io.vertx.core.net.NetClient;
 import io.vertx.core.net.NetClientOptions;
-import io.vertx.core.net.NetServer;
 import io.vertx.core.net.NetServerOptions;
 import io.vertx.core.net.SocketAddress;
 
@@ -78,7 +77,7 @@ public class DummyVertxMetrics implements VertxMetrics {
   }
 
   @Override
-  public TCPMetrics createMetrics(NetServer server, SocketAddress localAddress, NetServerOptions options) {
+  public TCPMetrics createMetrics(SocketAddress localAddress, NetServerOptions options) {
     return DummyTCPMetrics.INSTANCE;
   }
