+++ src/main/java/io/vertx/core/net/impl/NetServerImpl.java
@@ -16,36 +16,25 @@
 
 package io.vertx.core.net.impl;
 
-import io.netty.bootstrap.ServerBootstrap;
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.*;
-import io.netty.channel.group.ChannelGroup;
-import io.netty.channel.group.ChannelGroupFuture;
-import io.netty.channel.group.DefaultChannelGroup;
-import io.netty.channel.socket.nio.NioServerSocketChannel;
 import io.netty.handler.logging.LoggingHandler;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.handler.stream.ChunkedWriteHandler;
 import io.netty.handler.timeout.IdleStateHandler;
-import io.netty.util.concurrent.GlobalEventExecutor;
 import io.vertx.core.AsyncResult;
-import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.Closeable;
+import io.vertx.core.buffer.Buffer;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
 import io.vertx.core.net.NetServer;
 import io.vertx.core.net.NetServerOptions;
 import io.vertx.core.net.NetSocket;
 import io.vertx.core.net.NetSocketStream;
-import io.vertx.core.spi.metrics.Metrics;
-import io.vertx.core.spi.metrics.MetricsProvider;
 import io.vertx.core.spi.metrics.TCPMetrics;
 
-import java.net.InetSocketAddress;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
+import static io.vertx.core.net.impl.VertxHandler.safeBuffer;
 
 /**
  *
@@ -53,56 +42,60 @@ import java.util.concurrent.ConcurrentHashMap;
  *
  * @author <a href="http://tfox.org">Tim Fox</a>
  */
-public class NetServerImpl implements NetServer, Closeable, MetricsProvider {
+public class NetServerImpl extends NetServerBase<NetSocketImpl> implements NetServer {
 
-  private static final Logger log = LoggerFactory.getLogger(NetServerImpl.class);
-
-  private final VertxInternal vertx;
-  private final NetServerOptions options;
-  private final ContextImpl creatingContext;
-  private final SSLHelper sslHelper;
-  private final Map<Channel, NetSocketImpl> socketMap = new ConcurrentHashMap<>();
-  private final VertxEventLoopGroup availableWorkers = new VertxEventLoopGroup();
-  private final HandlerManager<Handler<NetSocket>> handlerManager = new HandlerManager<>(availableWorkers);
   private final NetSocketStreamImpl connectStream = new NetSocketStreamImpl();
-  private final boolean logEnabled;
-  private ChannelGroup serverChannelGroup;
-  private volatile boolean listening;
-  private volatile ServerID id;
-  private NetServerImpl actualServer;
-  private AsyncResolveConnectHelper bindFuture;
-  private volatile int actualPort;
-  private ContextImpl listenContext;
-  private TCPMetrics metrics;
+  private Handler<NetSocket> handler;
+  private Handler<Void> endHandler;
 
   public NetServerImpl(VertxInternal vertx, NetServerOptions options) {
-    this.vertx = vertx;
-    this.options = new NetServerOptions(options);
-    this.sslHelper = new SSLHelper(options, options.getKeyCertOptions(), options.getTrustOptions());
-    this.creatingContext = vertx.getContext();
-    this.logEnabled = options.getLogActivity();
-    if (creatingContext != null) {
-      if (creatingContext.isMultiThreadedWorkerContext()) {
-        throw new IllegalStateException("Cannot use NetServer in a multi-threaded worker verticle");
-      }
-      creatingContext.addCloseHook(this);
-    }
+    super(vertx, options);
   }
 
   @Override
-  public NetServer connectHandler(Handler<NetSocket> handler) {
-    connectStream.handler(handler);
+  public synchronized Handler<NetSocket> connectHandler() {
+    return handler;
+  }
+
+  @Override
+  public synchronized NetServer connectHandler(Handler<NetSocket> handler) {
+    if (isListening()) {
+      throw new IllegalStateException("Cannot set connectHandler when server is listening");
+    }
+    this.handler = handler;
     return this;
   }
 
   @Override
-  public Handler<NetSocket> connectHandler() {
-    return connectStream.handler();
+  protected void initChannel(ChannelPipeline pipeline) {
+    if (sslHelper.isSSL()) {
+      SslHandler sslHandler = sslHelper.createSslHandler(vertx);
+      pipeline.addLast("ssl", sslHandler);
+    }
+    if (logEnabled) {
+      pipeline.addLast("logging", new LoggingHandler());
+    }
+    if (sslHelper.isSSL()) {
+      // only add ChunkedWriteHandler when SSL is enabled otherwise it is not needed as FileRegion is used.
+      pipeline.addLast("chunkedWriter", new ChunkedWriteHandler());       // For large file / sendfile support
+    }
+    if (options.getIdleTimeout() > 0) {
+      pipeline.addLast("idle", new IdleStateHandler(0, 0, options.getIdleTimeout()));
+    }
   }
 
   @Override
-  public NetSocketStream connectStream() {
-    return connectStream;
+  protected void handleMsgReceived(NetSocketImpl conn, Object msg) {
+    ByteBuf buf = (ByteBuf) msg;
+    conn.handleDataReceived(Buffer.buffer(buf));
+  }
+
+  @Override
+  protected Object safeObject(Object msg, ByteBufAllocator allocator) {
+    if (msg instanceof ByteBuf) {
+      return safeBuffer((ByteBuf) msg, allocator);
+    }
+    return msg;
   }
 
   @Override
@@ -132,370 +125,68 @@ public class NetServerImpl implements NetServer, Closeable, MetricsProvider {
   }
 
   @Override
-  public synchronized NetServer listen(int port, String host, Handler<AsyncResult<NetServer>> listenHandler) {
-    if (connectStream.handler() == null) {
-      throw new IllegalStateException("Set connect handler first");
-    }
-    if (listening) {
-      throw new IllegalStateException("Listen already called");
-    }
-    listening = true;
-
-    listenContext = vertx.getOrCreateContext();
-
-    synchronized (vertx.sharedNetServers()) {
-      this.actualPort = port; // Will be updated on bind for a wildcard port
-      id = new ServerID(port, host);
-      NetServerImpl shared = vertx.sharedNetServers().get(id);
-      if (shared == null || port == 0) { // Wildcard port will imply a new actual server each time
-        serverChannelGroup = new DefaultChannelGroup("vertx-acceptor-channels", GlobalEventExecutor.INSTANCE);
-
-        ServerBootstrap bootstrap = new ServerBootstrap();
-        bootstrap.group(availableWorkers);
-        bootstrap.channel(NioServerSocketChannel.class);
-        sslHelper.validate(vertx);
-
-        bootstrap.childHandler(new ChannelInitializer<Channel>() {
-          @Override
-          protected void initChannel(Channel ch) throws Exception {
-            if (connectStream.isPaused()) {
-              ch.close();
-              return;
-            }
-            ChannelPipeline pipeline = ch.pipeline();
-            if (sslHelper.isSSL()) {
-              SslHandler sslHandler = sslHelper.createSslHandler(vertx);
-              pipeline.addLast("ssl", sslHandler);
-            }
-            if (logEnabled) {
-              pipeline.addLast("logging", new LoggingHandler());
-            }
-            if (sslHelper.isSSL()) {
-              // only add ChunkedWriteHandler when SSL is enabled otherwise it is not needed as FileRegion is used.
-              pipeline.addLast("chunkedWriter", new ChunkedWriteHandler());       // For large file / sendfile support
-            }
-            if (options.getIdleTimeout() > 0) {
-              pipeline.addLast("idle", new IdleStateHandler(0, 0, options.getIdleTimeout()));
-            }
-            pipeline.addLast("handler", new ServerHandler(ch));
-          }
-        });
-
-        applyConnectionOptions(bootstrap);
-
-        if (connectStream.handler() != null) {
-          handlerManager.addHandler(connectStream.handler(), listenContext);
-        }
-
-        try {
-          bindFuture = AsyncResolveConnectHelper.doBind(vertx, port, host, bootstrap);
-          bindFuture.addListener(res -> {
-            if (res.succeeded()) {
-              Channel ch = res.result();
-              log.trace("Net server listening on " + host + ":" + ch.localAddress());
-              // Update port to actual port - wildcard port 0 might have been used
-              NetServerImpl.this.actualPort = ((InetSocketAddress)ch.localAddress()).getPort();
-              NetServerImpl.this.id = new ServerID(NetServerImpl.this.actualPort, id.host);
-              serverChannelGroup.add(ch);
-              vertx.sharedNetServers().put(id, NetServerImpl.this);
-              metrics = vertx.metricsSPI().createMetrics(this, new SocketAddressImpl(id.port, id.host), options);
-            } else {
-              vertx.sharedNetServers().remove(id);
-            }
-          });
-
-        } catch (Throwable t) {
-          // Make sure we send the exception back through the handler (if any)
-          if (listenHandler != null) {
-            vertx.runOnContext(v ->  listenHandler.handle(Future.failedFuture(t)));
-          } else {
-            // No handler - log so user can see failure
-            log.error(t);
-          }
-          listening = false;
-          return this;
-        }
-        if (port != 0) {
-          vertx.sharedNetServers().put(id, this);
-        }
-        actualServer = this;
-      } else {
-        // Server already exists with that host/port - we will use that
-        actualServer = shared;
-        this.actualPort = shared.actualPort();
-        metrics = vertx.metricsSPI().createMetrics(this, new SocketAddressImpl(id.port, id.host), options);
-        if (connectStream.handler() != null) {
-          actualServer.handlerManager.addHandler(connectStream.handler(), listenContext);
-        }
-      }
-
-      // just add it to the future so it gets notified once the bind is complete
-      actualServer.bindFuture.addListener(res -> {
-        if (listenHandler != null) {
-          AsyncResult<NetServer> ares;
-          if (res.succeeded()) {
-            ares = Future.succeededFuture(NetServerImpl.this);
-          } else {
-            listening = false;
-            ares = Future.failedFuture(res.cause());
-          }
-          // Call with expectRightThread = false as if server is already listening
-          // Netty will call future handler immediately with calling thread
-          // which might be a non Vert.x thread (if running embedded)
-          listenContext.runOnContext(v -> listenHandler.handle(ares));
-        } else if (res.failed()) {
-          // No handler - log so user can see failure
-          log.error("Failed to listen", res.cause());
-          listening = false;
-        }
-      });
-    }
+  public synchronized NetServerImpl listen(int port, String host, Handler<AsyncResult<NetServer>> listenHandler) {
+    listen(handler, port, host, ar -> listenHandler.handle(ar.map(this)));
     return this;
   }
 
-  public synchronized void close() {
-    close(null);
+  @Override
+  public NetSocketStream connectStream() {
+    return connectStream;
   }
 
   @Override
   public synchronized void close(Handler<AsyncResult<Void>> done) {
-    if (connectStream.endHandler() != null) {
-      Handler<Void> endHandler = connectStream.endHandler;
-      connectStream.endHandler = null;
+    if (endHandler != null) {
+      Handler<Void> handler = endHandler;
+      endHandler = null;
       Handler<AsyncResult<Void>> next = done;
       done = event -> {
         if (event.succeeded()) {
-          endHandler.handle(event.result());
+          handler.handle(event.result());
         }
         if (next != null) {
           next.handle(event);
         }
       };
     }
-
-    ContextImpl context = vertx.getOrCreateContext();
-    if (!listening) {
-      if (done != null) {
-        executeCloseDone(context, done, null);
-      }
-      return;
-    }
-    listening = false;
-    synchronized (vertx.sharedNetServers()) {
-
-      if (actualServer != null) {
-        actualServer.handlerManager.removeHandler(connectStream.handler(), listenContext);
-
-        if (actualServer.handlerManager.hasHandlers()) {
-          // The actual server still has handlers so we don't actually close it
-          if (done != null) {
-            executeCloseDone(context, done, null);
-          }
-        } else {
-          // No Handlers left so close the actual server
-          // The done handler needs to be executed on the context that calls close, NOT the context
-          // of the actual server
-          actualServer.actualClose(context, done);
-        }
-      }
-    }
-    if (creatingContext != null) {
-      creatingContext.removeCloseHook(this);
-    }
+    super.close(done);
   }
 
   @Override
-  public synchronized int actualPort() {
-    return actualPort;
-  }
-
-  @Override
-  public boolean isMetricsEnabled() {
-    return metrics != null && metrics.isEnabled();
-  }
-
-  @Override
-  public Metrics getMetrics() {
-    return metrics;
-  }
-
-  private void applyConnectionOptions(ServerBootstrap bootstrap) {
-    bootstrap.childOption(ChannelOption.TCP_NODELAY, options.isTcpNoDelay());
-    if (options.getSendBufferSize() != -1) {
-      bootstrap.childOption(ChannelOption.SO_SNDBUF, options.getSendBufferSize());
-    }
-    if (options.getReceiveBufferSize() != -1) {
-      bootstrap.childOption(ChannelOption.SO_RCVBUF, options.getReceiveBufferSize());
-      bootstrap.childOption(ChannelOption.RCVBUF_ALLOCATOR, new FixedRecvByteBufAllocator(options.getReceiveBufferSize()));
-    }
-    if (options.getSoLinger() != -1) {
-      bootstrap.option(ChannelOption.SO_LINGER, options.getSoLinger());
-    }
-    if (options.getTrafficClass() != -1) {
-      bootstrap.childOption(ChannelOption.IP_TOS, options.getTrafficClass());
-    }
-    bootstrap.childOption(ChannelOption.ALLOCATOR, PartialPooledByteBufAllocator.INSTANCE);
-
-    bootstrap.childOption(ChannelOption.SO_KEEPALIVE, options.isTcpKeepAlive());
-    bootstrap.option(ChannelOption.SO_REUSEADDR, options.isReuseAddress());
-    if (options.getAcceptBacklog() != -1) {
-      bootstrap.option(ChannelOption.SO_BACKLOG, options.getAcceptBacklog());
-    }
-  }
-
-  private void actualClose(ContextImpl closeContext, Handler<AsyncResult<Void>> done) {
-    if (id != null) {
-      vertx.sharedNetServers().remove(id);
-    }
-
-    ContextImpl currCon = vertx.getContext();
-
-    for (NetSocketImpl sock : socketMap.values()) {
-      sock.close();
-    }
-
-    // Sanity check
-    if (vertx.getContext() != currCon) {
-      throw new IllegalStateException("Context was changed");
-    }
-
-    ChannelGroupFuture fut = serverChannelGroup.close();
-    fut.addListener(cg -> {
-      if (metrics != null) {
-        metrics.close();
-      }
-      executeCloseDone(closeContext, done, fut.cause());
-    });
-
-  }
-
-  private void executeCloseDone(ContextImpl closeContext, Handler<AsyncResult<Void>> done, Exception e) {
-    if (done != null) {
-      Future<Void> fut = e == null ? Future.succeededFuture() : Future.failedFuture(e);
-      closeContext.runOnContext(v -> done.handle(fut));
-    }
-  }
-
-  private class ServerHandler extends VertxNetHandler {
-    public ServerHandler(Channel ch) {
-      super(ch, socketMap);
-    }
-
-    @Override
-    public void channelActive(ChannelHandlerContext ctx) throws Exception {
-      Channel ch = ctx.channel();
-      EventLoop worker = ch.eventLoop();
-
-      //Choose a handler
-      HandlerHolder<Handler<NetSocket>> handler = handlerManager.chooseHandler(worker);
-      if (handler == null) {
-        //Ignore
-        return;
-      }
-
-      if (sslHelper.isSSL()) {
-        SslHandler sslHandler = ch.pipeline().get(SslHandler.class);
-
-        io.netty.util.concurrent.Future<Channel> fut = sslHandler.handshakeFuture();
-        fut.addListener(future -> {
-          if (future.isSuccess()) {
-            connected(ch, handler);
-          } else {
-            log.error("Client from origin " + ch.remoteAddress() + " failed to connect over ssl: " + future.cause());
-          }
-        });
-      } else {
-        connected(ch, handler);
-      }
-    }
-
-    private void connected(Channel ch, HandlerHolder<Handler<NetSocket>> handler) {
-      // Need to set context before constructor is called as writehandler registration needs this
-      ContextImpl.setContext(handler.context);
-      NetSocketImpl sock = new NetSocketImpl(vertx, ch, handler.context, sslHelper, metrics);
-      socketMap.put(ch, sock);
-      VertxNetHandler netHandler = ch.pipeline().get(VertxNetHandler.class);
-      netHandler.conn = sock;
-      handler.context.executeFromIO(() -> {
-        sock.metric(metrics.connected(sock.remoteAddress(), sock.remoteName()));
-        handler.handler.handle(sock);
-      });
-    }
-  }
-
-  @Override
-  protected void finalize() throws Throwable {
-    // Make sure this gets cleaned up if there are no more references to it
-    // so as not to leave connections and resources dangling until the system is shutdown
-    // which could make the JVM run out of file handles.
-    close();
-    super.finalize();
+  protected NetSocketImpl createConnection(VertxInternal vertx, Channel channel, ContextImpl context, SSLHelper helper, TCPMetrics metrics) {
+    return new NetSocketImpl(vertx, channel, context, helper, metrics);
   }
 
   /*
-    Needs to be protected using the NetServerImpl monitor as that protects the listening variable
-    In practice synchronized overhead should be close to zero assuming most access is from the same thread due
-    to biased locks
-  */
+        Needs to be protected using the NetServerImpl monitor as that protects the listening variable
+        In practice synchronized overhead should be close to zero assuming most access is from the same thread due
+        to biased locks
+      */
   private class NetSocketStreamImpl implements NetSocketStream {
-
-    private Handler<NetSocket> handler;
-    private boolean paused;
-    private Handler<Void> endHandler;
-
-    Handler<NetSocket> handler() {
-      synchronized (NetServerImpl.this) {
-        return handler;
-      }
-    }
-
-    boolean isPaused() {
-      synchronized (NetServerImpl.this) {
-        return paused;
-      }
-    }
-
-     Handler<Void> endHandler() {
-       synchronized (NetServerImpl.this) {
-         return endHandler;
-       }
-    }
 
     @Override
     public NetSocketStreamImpl handler(Handler<NetSocket> handler) {
-      synchronized (NetServerImpl.this) {
-        if (listening) {
-          throw new IllegalStateException("Cannot set connectHandler when server is listening");
-        }
-        this.handler = handler;
-        return this;
-      }
+      connectHandler(handler);
+      return this;
     }
 
     @Override
     public NetSocketStreamImpl pause() {
-      synchronized (NetServerImpl.this) {
-        if (!paused) {
-          paused = true;
-        }
-        return this;
-      }
+      pauseAccepting();
+      return this;
     }
 
     @Override
     public NetSocketStreamImpl resume() {
-      synchronized (NetServerImpl.this) {
-        if (paused) {
-          paused = false;
-        }
-        return this;
-      }
+      resumeAccepting();
+      return this;
     }
 
     @Override
-    public NetSocketStreamImpl endHandler(Handler<Void> endHandler) {
+    public NetSocketStreamImpl endHandler(Handler<Void> handler) {
       synchronized (NetServerImpl.this) {
-        this.endHandler = endHandler;
+        endHandler = handler;
         return this;
       }
     }
